Add files via upload
The Object Detection Based Point of Sale (POS) System is an intelligent billing solution designed to automate the product identification and checkout process in retail environments. This system leverages YOLOv8, a state-of-the-art real-time object detection algorithm, to accurately detect and classify products placed in front of a camera, and Flask, a lightweight Python web framework, to provide a user-friendly web interface and backend services.